requirements.txt

asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-dotenv==1.4.2
django-getenv==1.3.2
djangorestframework==3.14.0
idna==3.4
pip==23.0.1
psycopg2==2.9.7
pytz==2023.3
requests==2.31.0
setuptools==66.1.1
sqlparse==0.4.4
urllib3==2.0.4